requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.3
fake-useragent>=2.2.0
platformdirs>=4.4.0
socksio>=1.0.0
PyYAML~=6.0.2
streamlit~=1.49.1
pandas~=2.3.2
psutil~=7.1.0
